extract ReportRouteParamHandler from ReportScreen
